[CHANGED] Split ZeebeAdapter into four classes, ZeebeAdapter is now a mixin [ADDED] many tests for ZeebeAdapters [Added] message_id param to publish_message on ZeebeClient and ZeebeAdapter
