+++ pyzeebe/grpc_internals/zeebe_adapter_base.py
@@ -1,0 +1,79 @@
+import logging
+import os
+import os.path
+
+import grpc
+
+from pyzeebe.common.exceptions import *
+from pyzeebe.credentials.base_credentials import BaseCredentials
+from pyzeebe.grpc_internals.zeebe_pb2_grpc import GatewayStub
+
+
+class ZeebeAdapterBase(object):
+    def __init__(self, hostname: str = None, port: int = None, credentials: BaseCredentials = None,
+                 channel: grpc.Channel = None, secure_connection: bool = False):
+        if channel:
+            self.connection_uri = None
+            self._channel = channel
+        else:
+            self.connection_uri = self._get_connection_uri(hostname, port, credentials)
+            self._channel = self._create_channel(self.connection_uri, credentials, secure_connection)
+
+        self.secure_connection = secure_connection
+        self.connected = False
+        self.retrying_connection = True
+        self._channel.subscribe(self._check_connectivity, try_to_connect=True)
+        self.gateway_stub = GatewayStub(self._channel)
+
+    @staticmethod
+    def _get_connection_uri(hostname: str = None, port: int = None, credentials: BaseCredentials = None) -> str:
+        if credentials and credentials.get_connection_uri():
+            return credentials.get_connection_uri()
+        if hostname or port:
+            return f"{hostname or 'localhost'}:{port or 26500}"
+        else:
+            return os.getenv("ZEEBE_ADDRESS", "localhost:26500")
+
+    @staticmethod
+    def _create_channel(connection_uri: str, credentials: BaseCredentials = None,
+                        secure_connection: bool = False) -> grpc.Channel:
+        if credentials:
+            return grpc.secure_channel(connection_uri, credentials.grpc_credentials)
+        elif secure_connection:
+            return grpc.secure_channel(connection_uri, grpc.ssl_channel_credentials())
+        else:
+            return grpc.insecure_channel(connection_uri)
+
+    def _check_connectivity(self, value: grpc.ChannelConnectivity) -> None:
+        logging.debug(f"Grpc channel connectivity changed to: {value}")
+        if value in [grpc.ChannelConnectivity.READY, grpc.ChannelConnectivity.IDLE]:
+            logging.debug(f"Connected to {self.connection_uri or 'zeebe'}")
+            self.connected = True
+            self.retrying_connection = False
+        elif value == grpc.ChannelConnectivity.CONNECTING:
+            logging.debug(f"Connecting to {self.connection_uri or 'zeebe'}.")
+            self.connected = False
+            self.retrying_connection = True
+        elif value == grpc.ChannelConnectivity.TRANSIENT_FAILURE:
+            logging.warning(f"Lost connection to {self.connection_uri or 'zeebe'}. Retrying...")
+            self.connected = False
+            self.retrying_connection = True
+        elif value == grpc.ChannelConnectivity.SHUTDOWN:
+            logging.error(f"Failed to establish connection to {self.connection_uri or 'zeebe'}. Non recoverable")
+            self.connected = False
+            self.retrying_connection = False
+            raise ConnectionAbortedError(f"Lost connection to {self.connection_uri or 'zeebe'}")
+
+    def _common_zeebe_grpc_errors(self, rpc_error: grpc.RpcError):
+        if self.is_error_status(rpc_error, grpc.StatusCode.RESOURCE_EXHAUSTED):
+            raise ZeebeBackPressure()
+        elif self.is_error_status(rpc_error, grpc.StatusCode.UNAVAILABLE):
+            raise ZeebeGatewayUnavailable()
+        elif self.is_error_status(rpc_error, grpc.StatusCode.INTERNAL):
+            raise ZeebeInternalError()
+        else:
+            raise rpc_error
+
+    @staticmethod
+    def is_error_status(rpc_error: grpc.RpcError, status_code: grpc.StatusCode):
+        return rpc_error._state.code == status_code
