+++ pyzeebe/grpc_internals/zeebe_message_adapter.py
@@ -1,0 +1,22 @@
+import json
+from typing import Dict
+
+import grpc
+
+from pyzeebe.common.exceptions import *
+from pyzeebe.grpc_internals.zeebe_adapter_base import ZeebeAdapterBase
+from pyzeebe.grpc_internals.zeebe_pb2 import *
+
+
+class ZeebeMessageAdapter(ZeebeAdapterBase):
+    def publish_message(self, name: str, correlation_key: str, time_to_live_in_milliseconds: int,
+                        variables: Dict, message_id: str = None) -> PublishMessageResponse:
+        try:
+            return self.gateway_stub.PublishMessage(
+                PublishMessageRequest(name=name, correlationKey=correlation_key, messageId=message_id,
+                                      timeToLive=time_to_live_in_milliseconds, variables=json.dumps(variables)))
+        except grpc.RpcError as rpc_error:
+            if self.is_error_status(rpc_error, grpc.StatusCode.ALREADY_EXISTS):
+                raise MessageAlreadyExists()
+            else:
+                self._common_zeebe_grpc_errors(rpc_error)
