+++ tests/unit/utils/grpc_utils.py
@@ -1,0 +1,6 @@
+import grpc
+
+
+class GRPCStatusCode:
+    def __init__(self, code: grpc.StatusCode):
+        self.code = code
